Move Pytorch collators to separate module
